Slowly Re-Integrating my code from the latest version of ReplicatorG. This is easier than doing a massive merge with conficts.
Second commit. Adding pausegraph function.

All done! Just need to get rid of the 32bit eeprom write command and test this out on a makerbot!

Changed a few lines of code to make it a bit easier to understand what I am calling. (Commands)
Also switched over to using pin D9 for the ZProbe. D10 has been disabled in the firmware to allow for external steppers on the Extruder controller.

I have tested out all of this code on a Thingomatic with a DC motor extruder and am happy to say it works! However, someone should test this on a Thingomatic with a Stepper extruder (JustInCase).